style
update style css more modern
