+++ src/static_html/404.html
@@ -24,9 +24,9 @@
       <div class="header-container">
         <div class="header">
           <div class="container-logo-title">
-            <h1 class="header-title">L'ets go out to Paris</h1>
+            <h1 class="header-title">L'ets go out in Paris</h1>
             <h2 class="title-description">
-              Explorez et partagez les meilleurs endroits de Paris
+              Explorez et partagez les meilleurs endroits dans paris
             </h2>
           </div>
         
@@ -81,7 +81,7 @@
           </div>
 
           <p class="mention-legale">
-            &copy; 2023 L'ets go out to Paris. Tous droits réservés.
+            &copy; 2023 L'ets go out in Paris. Tous droits réservés.
           </p>
         </div>
       </div>
