+++ src/static_html/admin-modified-review.html
@@ -25,7 +25,7 @@
         <div class="header">
           <div class="container-logo-title">
             <a href="../static_html/main-review.html">
-              <h1 class="header-title-admin">L'ets go out to Paris</h1>
+              <h1 class="header-title-admin">L'ets go out in Paris</h1>
             </a>
             <h2 class="title-description">
               Administrateur - Modifications en attente
