+++ src/static_html/home-page.html
@@ -8,12 +8,12 @@
 
     <meta
       name="description"
-      content="Découvrez Paris comme jamais auparavant avec Let's Go Out to Paris. Explorez les trésors cachés et les joyaux de la Ville Lumière de manière captivante."
+      content="Découvrez Paris comme jamais auparavant avec Let's Go Out in Paris. Explorez les trésors cachés et les joyaux de la Ville Lumière de manière captivante."
     />
 
   
     <title>
-      Découvrez les meilleurs endroits à Paris avec Let's Go Out to Paris
+      Découvrez les meilleurs endroits à Paris avec Let's Go Out in Paris
     </title>
 
     <!--ci dessous permet l'utilisation de la font 'courgette' (google font)-->
@@ -41,14 +41,14 @@
 
         <header>
           <p class="bienvenu">bienvenu sur </p>
-          <h1 class="home-title"> Let's Go Out to Paris</h1>
+          <h1 class="home-title"> Let's Go Out in Paris</h1>
         </header>
 
         <article>
           <p class="home-description">
             Explorez la Ville Lumière comme jamais auparavant avec notre
             plateforme interactive qui vous révèle les trésors cachés et les
-            joyaux de Paris. "Let's Go Out to Paris" vous offre une expérience
+            joyaux dans paris. "Let's Go Out in Paris" vous offre une expérience
             unique pour découvrir la vie parisienne d'une manière captivante.
           </p>
