+++ src/static_html/main-review.html
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="../static_style/static_mobile.css" />
    
    
-   <title>L'ets go out to Paris -  Découvrez les Meilleurs Endroits de Paris</title>
+   <title>L'ets go out in Paris -  Découvrez les Meilleurs Endroits dans paris</title>
     <meta name="description" content="LGOP vous propose des annonces mêlant loisirs, culture, et food & drink pour assurer le succès de votre prochaine sortie à Paris. Découvrez les trésors cachés de la Ville Lumière avec notre plateforme unique." />
  
 
@@ -29,9 +29,9 @@
           <div class="container-logo-title">
           
 
-            <h1 class="header-title">L'ets go out to Paris</h1>
+            <h1 class="header-title">L'ets go out in Paris</h1>
 
-            <h2 class="title-description"> Explorez et partagez les meilleurs endroits de Paris</h2>
+            <h2 class="title-description"> Explorez et partagez les meilleurs endroits dans paris</h2>
           </div>
           <div class="user-connexion-mobile">
             <img class="user-logo" src="../assets/user.png" alt="user" />
@@ -958,7 +958,7 @@
           </div>
 
           <p class="mention-legale">
-            &copy; 2023 L'ets go out to Paris. Tous droits réservés.
+            &copy; 2023 L'ets go out in Paris. Tous droits réservés.
           </p>
         </div>
       </div>
