+++ src/static_html/signup.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="../static_style/static_mobile.css" />
 
     <title>
-      L'ets go out to Paris - Découvrez les Meilleurs Endroits de Paris
+      L'ets go out in Paris - Découvrez les Meilleurs Endroits dans paris
     </title>
 
     <meta
@@ -30,10 +30,10 @@
         <div class="header">
           <div class="container-logo-title">
             <a href="../static_html/main-review.html">
-              <h1 class="header-title">Let's go out to Paris</h1>
+              <h1 class="header-title">Let's go out in Paris</h1>
             </a>
             <h2 class="title-description">
-              Explorez et partagez les meilleurs endroits de Paris
+              Explorez et partagez les meilleurs endroits dans paris
             </h2>
           </div>
           <div class="user-connexion-mobile">
@@ -246,7 +246,7 @@
           </div>
 
           <p class="mention-legale">
-            &copy; 2023 L'ets go out to Paris. Tous droits réservés.
+            &copy; 2023 L'ets go out in Paris. Tous droits réservés.
           </p>
         </div>
       </div>
