+++ src/styles/resonsive-desktop.css
@@ -17,8 +17,7 @@
     grid-template-columns: repeat(4, 1fr);
     /* 4 => nb colonnes ; 1fr element penant 1 fraction égale de l'espace disponible. */
 
-    margin: 4rem 2rem 0 2rem;  
-  
+    margin: 5rem 2rem 5.5rem 2rem;  
 
   }
 
@@ -30,13 +29,12 @@
   }
 
 
+
   .card-container:hover {
+background :#afe2e2;
 
-    background: linear-gradient(to right, #c09ab0, #b3a6b0, #a18e9a, #d1c3cb);
-    
 
-  }
-
+ }
 
   .card-container:hover::before {
 
@@ -61,7 +59,7 @@
     height: 1.6rem;
     align-items: flex-end;
     margin-right: 12rem;
-    margin-top: 8.4rem;
+    margin-top: 11rem;
     color: #462243;
     font-family: "Courgette";
     font-size: auto;
@@ -78,40 +76,40 @@
   
 
   button.create-review {
+
     position: absolute;
     right: 10rem;
-    font-family: "Courgette";
-    color: white;
-    background-color: rgb(137, 120, 148);
+    color: rgb(68, 50, 71);
+    background-color: rgb(210, 203, 218);
     border-radius: 0 0 10px 10px;
     width: auto;
     height: auto;
     margin-bottom: 5rem;
     font-size: 1.4rem;
+    font-family: 'Times New Roman', Times, serif;
     box-shadow: 0px 4px 8px rgba(14, 2, 19, 0.5);
+    outline: none;
+
   }
 
   button.create-review-after-scrollY {
     
     position: fixed;
     height: 2rem;
-    background-color:rgb(137, 120, 148);
-    color:white;
+    background-color: rgb(210, 203, 218);
+    color:rgb(122, 109, 128);
     z-index: 1;
-    right: -4.2rem;
-    top:8rem;
-    font-family: "Courgette";
+    right:-4.2rem;
+    top: 8rem;
+    font-family: 'Times New Roman', Times, serif;
     border-radius: 0 0 15px 15px;
     vertical-align: top;
-    font-size: 1.3rem;
-
+    font-size: 1.5rem;
     transform: rotate(7deg);
     -ms-transform: rotate(90deg);/* IE 9 */
     -moz-transform: rotate(90deg);/* Firefox */
     -webkit-transform: rotate(90deg);/* Safari and Chrome */
     -o-transform: rotate(90deg);/* Opera */
-    
-
 
   }
